requirements.txt
Flask>=2.0
google-cloud-firestore>=2.5
jsonschema>=4.0
gunicorn>=20.0 # For production server in Cloud Run